requirements.txt
aniso8601==9.0.1
attrs==22.2.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
confection==0.0.4
cryptography==39.0.1
cymem==2.0.7
document==1.0
docx2txt==0.8
en-core-web-sm==3.4.1
Flask==2.2.2
flask-restx==1.0.5
flask-swagger==0.2.14
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonify==0.5
jsonschema==4.17.3
langcodes==3.3.0
lxml==4.9.2
MarkupSafe==2.1.2
murmurhash==1.0.9
nltk==3.4.5
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pathy==0.10.1
pdfminer.six==20221105
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
plac==1.1.3
preshed==3.0.8
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.4
pypdf2==3.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
six==1.16.0
smart-open==6.3.0
sortedcontainers==2.4.0
spacy==3.4.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.5
thinc==8.1.7
tqdm==4.64.1
typer==0.7.0
typing-extensions==4.4.0
urllib3==1.26.14
wasabi==0.10.1
Werkzeug==2.2.2
zipp==3.12.1
